% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{data}
\name{bayescomp_colors}
\alias{bayescomp_colors}
\alias{bc_colors}
\alias{bc_colors_discrete}
\title{bayescomp color palettes}
\format{
Named character vectors.

An object of class \code{character} of length 8.

An object of class \code{character} of length 8.
}
\usage{
bc_colors

bc_colors_discrete
}
\description{
Named color palettes for benchmark comparison plots. The \code{bayescomp}
palette is designed for colorblind accessibility.
}
\keyword{datasets}
