% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_color_bayescomp}
\alias{scale_color_bayescomp}
\alias{scale_colour_bayescomp}
\title{Discrete color scale for bayescomp}
\usage{
scale_color_bayescomp(..., reverse = FALSE)

scale_colour_bayescomp(..., reverse = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{reverse}{Logical. Reverse palette order?}
}
\value{
A ggplot2 scale.
}
\description{
Discrete color scale for bayescomp
}
\examples{
library(ggplot2)
data <- bc_simulate_benchmark(n_systems = 10, n_methods = 3)
ggplot(data, aes(system_id, count, color = method)) +
  geom_point() + scale_color_bayescomp()
}
