% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{scale_fill_bayescomp}
\alias{scale_fill_bayescomp}
\title{Fill scale for bayescomp}
\usage{
scale_fill_bayescomp(..., reverse = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}

\item{reverse}{Logical. Reverse palette order?}
}
\value{
A ggplot2 scale.
}
\description{
Fill scale for bayescomp
}
\examples{
library(ggplot2)
data <- bc_simulate_benchmark(n_systems = 10, n_methods = 3)
ggplot(data, aes(method, count, fill = method)) +
  geom_boxplot() + scale_fill_bayescomp()
}
